minor refactor, added code segmentation inside data models and data collections
