+++ TT_2_Functional_Programming/Functional Programming/Assets/Scripts/Base/Data/DataCollection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e689bd4bc8ef4cfdb06949fecb83b3b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
