+++ TT_2_Functional_Programming/Functional Programming/Assets/Scripts/Base/Data/IDataCollection.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using System;
+using System.Linq;
+using Realtime.Messaging.Internal;
+using System.Text;
+
+public interface IDataCollection 
+{
+
+    List<Card> Data();
+}
