+++ TT_2_Functional_Programming/Functional Programming/Assets/Scripts/Base/Data/IDataCollection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25c471c8af4784aca92d6722b5d4c45b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
