+++ TT_2_Functional_Programming/Functional Programming/Assets/Scripts/Base/Data/IInteractor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32d31a7b0d060430ba01686e13f73856
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
