+++ TT_2_Functional_Programming/Functional Programming/Assets/Scripts/Base/Data/Interactor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79d2fc83544d54c3cb4a4e6782408638
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
